Updated the 0-1 Knapsack using DP
